% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotligraph.R
\name{plotligraph}
\alias{plotligraph}
\title{Manipulate \pkg{plotly} into making a 3D render of an \pkg{igraph} graph.}
\usage{
plotligraph(graph_obj, layout = NA, ...)
}
\arguments{
\item{graph_obj}{`igraph` object}

\item{layout}{optional `igraph` layout matrix, preferably in 3 dimensions.
Will generate with `layout.fruchterman.reingold` and dim = 3 otherwise.}

\item{...}{additional arguments passed to `plot_ly`.}
}
\value{
plot_ly object
}
\description{
Works by drawing a scatter3d of the vertices and tracing each edge with its
own scatter3d. Procs one RColorBrewer palette warning per edge, but don't
worry about it.
}
\examples{
library("plotly")
library("igraph")
library("plotligraph")

set.seed(1234)
graph <- barabasi.game(35)
E(graph)$name <- sample(letters, length(E(graph)), TRUE)
plotligraph(graph)
}

